+++ src/App.tsx
@@ -1,6 +1,8 @@
 import { Home } from './pages/Home';
+import { useNotifications } from './hooks/useNotifications';
 
 function App() {
+  useNotifications();
   return <Home />;
 }